4960: Move ingest batcher store back to single module
