+++ .github/workflows/chunk-clients-cdk.yaml
@@ -28,7 +28,6 @@ on:
       - 'java/ingest/ingest-trackerV2/**'
       - 'java/statestorev2/**'
       - 'java/ingest/ingest-batcher-store/**'
-      - 'java/ingest/ingest-batcher-storev2/**'
       - 'java/common/common-task/**'
       - 'java/common/common-taskv2/**'
       - 'java/query/query-runnerv2/**'
