+++ .github/workflows/chunk-ingest.yaml
@@ -16,7 +16,6 @@ on:
       - 'java/ingest/ingest-taskrunner/**'
       - 'java/ingest/ingest-batcher-core/**'
       - 'java/ingest/ingest-batcher-store/**'
-      - 'java/ingest/ingest-batcher-storev2/**'
       - 'java/ingest/ingest-batcher-submitter/**'
       - 'java/ingest/ingest-batcher-submitterv2/**'
       - 'java/ingest/ingest-batcher-job-creator/**'
