+++ java/clients/pom.xml
@@ -138,7 +138,7 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
-            <artifactId>ingest-batcher-storev2</artifactId>
+            <artifactId>ingest-batcher-store</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
