+++ java/clients/src/main/java/sleeper/clients/report/IngestBatcherReport.java
@@ -30,7 +30,7 @@ import sleeper.configurationv2.table.index.DynamoDBTableIndex;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.table.TableStatusProvider;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
-import sleeper.ingest.batcher.storev2.DynamoDBIngestBatcherStore;
+import sleeper.ingest.batcher.store.DynamoDBIngestBatcherStore;
 
 import java.util.HashMap;
 import java.util.Locale;
