+++ java/ingest/ingest-batcher-store/pom.xml
@@ -30,7 +30,7 @@
     <dependencies>
         <dependency>
             <groupId>sleeper</groupId>
-            <artifactId>dynamodb-tools</artifactId>
+            <artifactId>dynamodb-toolsv2</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
