+++ java/ingest/ingest-batcher-submitterv2/src/main/java/sleeper/ingest/batcher/submitterv2/IngestBatcherSubmitterLambda.java
@@ -35,7 +35,7 @@ import sleeper.core.table.TableIndex;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
 import sleeper.ingest.batcher.core.IngestBatcherSubmitRequest;
 import sleeper.ingest.batcher.core.IngestBatcherSubmitRequestSerDe;
-import sleeper.ingest.batcher.storev2.DynamoDBIngestBatcherStore;
+import sleeper.ingest.batcher.store.DynamoDBIngestBatcherStore;
 
 import java.time.Instant;
