+++ java/ingest/pom.xml
@@ -33,8 +33,6 @@
         <module>ingest-tracker</module>
         <module>ingest-trackerV2</module>
         <module>ingest-batcher-core</module>
-        <module>ingest-batcher-store</module>
-        <module>ingest-batcher-storev2</module>
         <module>ingest-batcher-submitter</module>
         <module>ingest-batcher-submitterv2</module>
         <module>ingest-batcher-job-creator</module>
